Installing memcached updating phalcon installer script and adding a sample html page

+++ vagrant/puphpet/files/exec-once/install_phalconphp.sh
@@ -1,8 +1,10 @@
-git clone git://github.com/phalcon/cphalcon.git
-cd cphalcon/build
-./install
-cd ..
-cd ..
-rm -rf cphalcon
-echo "extension=phalcon.so" > /etc/php.d/20-phalcon.ini
-service httpd restart
+if [ ! -f /etc/php.d/20-phalcon.ini ]; then
+    git clone git://github.com/phalcon/cphalcon.git
+    cd cphalcon/build
+    ./install
+    cd ..
+    cd ..
+    rm -rf cphalcon
+    echo "extension=phalcon.so" > /etc/php.d/20-phalcon.ini
+    service httpd restart
+fi
